requirements.txt
tensorflow==2.0.0-rc1
numpy
matplotlib
opencv-python
pillow
scikit-image
scipy